06_customer_support/disections/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
coloredlogs==15.0.1
concurrent-log-handler==0.9.25
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
deprecated==1.2.18
distro==1.9.0
durationpy==0.9
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
gradio==4.0.0
gradio-client==0.7.0
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
humanfriendly==10.0
idna==3.10
importlib-metadata==8.6.1
importlib-resources==6.5.2
ipython==8.34.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.22
langchain-community==0.3.20
langchain-core==0.3.50
langchain-openai==0.3.12
langchain-text-splitters==0.3.7
langgraph==0.3.24
langgraph-checkpoint==2.0.24
langgraph-checkpoint-postgres==2.0.19
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.23
markdown-it-py==3.0.0
markupsafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
narwhals==1.33.0
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.70.0
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.16
ormsgpack==1.9.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdfminer-six==20250327
pexpect==4.9.0
pillow==10.4.0
portalocker==3.1.1
posthog==3.23.0
prompt-toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic-core==2.33.1
pydantic-settings==2.8.1
pydub==0.25.1
pygments==2.19.1
pyparsing==3.2.3
pypika==0.48.9
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.40
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.3
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.12.0
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing-extensions==4.13.1
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
